Added redirect to landing page (#105)
