+++ components/Navbar/Navbar.tsx
@@ -10,6 +10,7 @@ import {
   logNavbarOpen,
   logNavbarNavigation,
 } from "../../api/Mixpanel.api";
+import { HOME_PAGE_URL } from '../../domain/Navigation';
 
 
 interface NavBarProps {
@@ -76,7 +77,7 @@ export const Navbar: React.FC<NavBarProps> = (props) => {
             width: navbarWidth,
           }}
         >
-          <MenuItem url="/" title="ראשי" />
+          <MenuItem url={HOME_PAGE_URL} title="ראשי" />
           <MenuItem url="/programs" title="תכניות" />
           <MenuItem url="/join-us" title="הצטרפות" />
           <MenuItem url="/about" title="הסיפור שלנו" />
@@ -95,7 +96,7 @@ export const Navbar: React.FC<NavBarProps> = (props) => {
         >
           <div className={styles.navbarHead}>
             <div className={styles.logo}>
-              <Link href="/">
+              <Link href={HOME_PAGE_URL}>
                 <img
                   src={`${BASE_IMAGE_ICON}radiosavta/logo_head`}
                   width="100%"
@@ -115,7 +116,7 @@ export const Navbar: React.FC<NavBarProps> = (props) => {
               <span></span>
             </div>
           </div>
-          <MenuItem url="/" title="ראשי" />
+          <MenuItem url={HOME_PAGE_URL} title="ראשי" />
           <MenuItem url="/programs" title="תכניות" />
           <MenuItem url="/join-us" title="הצטרפות" />
           <MenuItem url="/about" title="הסיפור שלנו" />
