+++ domain/Navigation.ts
@@ -1,0 +1,1 @@
+export const HOME_PAGE_URL = '/home';
